Introduce React\Socket\RawServer to connect to any stream supported by stream_socket_server
