+++ src/React/Socket/ServerInterface.php
@@ -7,7 +7,7 @@ use Evenement\EventEmitterInterface;
 /** @event connection */
 interface ServerInterface extends EventEmitterInterface
 {
-    public function listen($port, $host = '127.0.0.1');
+    public function listen($socket);
     public function getPort();
     public function shutdown();
 }
